feat: Anadir creacion de tableros de manera local y entrar al detalle del tablero

+++ server/src/main/kotlin/dev/wdona/burnt_out/Application.kt
@@ -29,7 +29,7 @@ fun Application.module() {
             val recibido = call.receive<Tarea>()
 
             val respuesta = Tarea(
-                idTarea = System.currentTimeMillis().toInt(),
+                idTarea = System.currentTimeMillis(),
                 titulo = recibido.titulo,
                 descripcion = recibido.descripcion,
                 estado = "pendiente",
